Bring over ext2/3/4 fixes:
ext2: Change inode from an int to long. Add support for creating an ext2 file from an inode but without an Ext2Entry.
ext4: Add support for reading file systems created with the flex_bg feature.
ext4: Detect multi-mount protection file systems. Add some tests for large directories and MMP.

+++ fs/src/fs/org/jnode/fs/ext2/Ext2Directory.java
@@ -213,7 +213,7 @@ public class Ext2Directory extends AbstractFSDirectory implements FSDirectoryId 
         synchronized (((Ext2FileSystem) getFileSystem()).getInodeCache()) {
             //reread the inode before synchronizing to it to make sure
             //all threads use the same instance
-            int iNodeNr = iNode.getINodeNr();
+            long iNodeNr = iNode.getINodeNr();
             iNode = ((Ext2FileSystem) getFileSystem()).getINode(iNodeNr);
 
             //lock the inode into the cache so it is not flushed before synchronizing to it
@@ -332,7 +332,7 @@ public class Ext2Directory extends AbstractFSDirectory implements FSDirectoryId 
 
     @Override
     public String getDirectoryId() {
-        return Integer.toString(iNode.getINodeNr());
+        return Long.toString(iNode.getINodeNr());
     }
 
     @Override
